#26 Database Service splitted
